CoreCLR projects don't need to be built first
Added the ability for the CoreCLR project in EDGE_APP_DIR to be
dynamically compiled at runtime instead of needing to be built first and
then requiring its assembly file to be referenced via a path.  So, if
EDGE_APP_DIR points to c:\Projects\MyApp, methods within the project can
be referenced via:

edge.func({
assemblyFile: 'MyApp',
typeName: 'MyApp.MyType',
methodName: 'MyMethod'
})

This negates the need to always build the app first and then point the
assemblyFile option to the build artifact.