Compilation fix for Linux w/ no Mono
When we build Edge.js on a Linux or OS X system without Mono installed,
we don't generate edge_nativeclr.node at all and when we load Edge at
runtime, we then default to trying to load edge_coreclr.node.  This
removes the need to specify the EDGE_USE_CORECLR environment variable
when using Edge on a Linux/OS X system without Mono.