make edge builds on node-chakracore
Piggy back on "tools/build.bat" to build for the current installed node
release when no VERSION is specified. If so save the build to subdir
node-ENGINE-VERSION instead of VERSION.

Update edge.js to check node-ENGINE-VERSION subdir before looking at known
versionMap.

Fix "whereis.js" where it was returnning "null" when target is not found.
```binding.gyp``` was testing ```console.log(result)``` output to be empty
string "" when target not found.