Error on load failures for explicit CLR versions
When an environment variable is used (either CORECLR_DIR or
CORECLR_VERSION) to specify a particular CLR target, we will now throw
an error if we fail to load the CLR from those targets.  This is in
contrast to previously where we would simply try to load it from the
next available target.