Avoid trying to load native code multiple times
When you depend on a module that requires edge and then you require edge
yourself, edge_coreclr.node or edge_nativeclr.node will be loaded and
initialized multiple times.  For edge_coreclr, the CLR will try to load
itself again which will, obviously, result in an error.  So, we set the
EDGE_NATIVE environment variable so that subsequent loads of the edge
module will try to load the native code from the same location and
therefore return the cached copy.