Use Nan::MakeCallback to avoid random hangs
Calling v8::Function::Call directly for async callbacks leads to very
odd behavior in some circumstances, specifically the next tick of the
event loop seems to get stuck. After much digging, I found references to
this being a "known" issue, and the proper way to call async functions
in JS is to use the node::MakeCallback helper, or the Nan equivalent. In
some cases, we also use the Nan::Callback object as a helper for
managing JS callbacks.

KNOWN ISSUES: On node <v6, it seems that exceptions thrown from JS when
called from .NET do NOT propagate back to .NET correctly. Instead, they
terminate the JS process. This appears to be a known limitation of using
the MakeCallback method, with no easy workarounds. On 6.4.0, this test
does pass, which indicates the issue has been fixed in node.