Explicitly kick the next tick instead of Nan::MakeCallback
The previous fixes to use Nan::MakeCallback had the side effect of
swallowing exceptions on Node <v6, which was a regression. To retain
that behavior, we revert the JS execution calls to using Nan::Call
instead of Nan::MakeCallback, and instead explicitly kick the next tick
with a no-op function.